Homologacao

+++ src/main/java/com/project/extension/controller/pedido/produto/dto/ProdutoRequestDto.java
@@ -7,13 +7,14 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ProdutoRequestDto(
              @NotBlank String nome,
              @NotBlank String descricao,
              @NotBlank String unidademedida,
-        @Positive Double preco,
+        @Positive BigDecimal preco,
              @NotNull Boolean ativo,
              @Valid @NotNull MetricaEstoqueRequestDto metrica,
              @Valid @NotNull List<AtributoProdutoRequestDto> atributos
