+++ src/main/java/com/project/extension/controller/pedido/servico/dto/servico/ServicoRequestDto.java
@@ -2,10 +2,12 @@ package com.project.extension.controller.pedido.servico.dto.servico;
 
 import jakarta.validation.constraints.NotBlank;
 
+import java.math.BigDecimal;
+
 public record ServicoRequestDto(
         @NotBlank String nome,
         @NotBlank String descricao,
-        Double precoBase,
+        BigDecimal precoBase,
         Boolean ativo,
         String etapaNome
 ) {
