+++ src/main/java/com/project/extension/controller/pedido/servico/dto/servico/ServicoResponseDto.java
@@ -3,6 +3,7 @@ package com.project.extension.controller.pedido.servico.dto.servico;
 import com.project.extension.controller.valueobject.etapa.EtapaResponseDto;
 import com.project.extension.controller.pedido.servico.dto.servico.agendamento.AgendamentoServicoResponseDto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,7 +13,7 @@ public record ServicoResponseDto(
         String codigo,
         String nome,
         String descricao,
-        Double precoBase,
+        BigDecimal precoBase,
         Boolean ativo,
         LocalDateTime createdAt,
         EtapaResponseDto etapa,
